+++ README.md
@@ -1,5 +1,7 @@
-# roc-plugin-style-less
-[![Build Status](https://travis-ci.org/rocjs/roc-plugin-style-less.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-style-less)
+# roc-plugin-style-less [![Build Status](https://travis-ci.org/rocjs/roc-plugin-style-less.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-style-less)
 
 __Plugin providing Less support for Webpack__  
-- [roc-plugin-style-less](/packages/roc-plugin-style-less)
+- [roc-plugin-style-less](/extensions/roc-plugin-style-less)
+
+---
+To be used together with [Roc](https://github.com/rocjs/roc).
