+++ extensions/roc-plugin-style-less/README.md
@@ -1,0 +1,15 @@
+# roc-plugin-style-less
+
+Plugin providing Less support for Webpack in Roc
+
+## Documentation
+- [Actions](/docs/Actions)
+- [Commands](/docs/Commands)
+- [Configuration](/docs/Configuration)
+- [Dependencies](/docs/Dependencies)
+- [Hooks](/docs/Hooks)
+- [Settings](/docs/Settings)
+- [Extensions](/docs/Extensions)
+
+---
+_Generated by [Roc](https://github.com/rocjs/roc)_
