+++ extensions/roc-plugin-style-less/docs/Settings.md
@@ -1,0 +1,3 @@
+# Settings for `roc-plugin-style-less`
+
+__No settings available.__
