+++ extensions/roc-plugin-style-less/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "roc-plugin-style-less",
+  "version": "1.0.0-beta.0",
+  "description": "Plugin providing Less support for Webpack in Roc",
+  "main": "lib/index.js",
+  "scripts": {
+    "lint": "eslint .",
+    "test": "npm run lint"
+  },
+  "files": [
+    "lib"
+  ],
+  "keywords": [
+    "roc",
+    "roc-plugin",
+    "webpack",
+    "style",
+    "less"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/rocjs/roc-plugin-style-less.git"
+  },
+  "author": "Aftonbladet AB",
+  "license": "MIT",
+  "repository": "https://github.com/rocjs/roc-plugin-style-less/tree/master/extensions/roc-plugin-style-less",
+  "author": "Aftonbladet AB",
+  "contributors": [{
+      "name": "Christian Lizell"
+    }, {
+      "name": "Gustaf Dalemar"
+  }],
+  "license": "MIT",
+  "dependencies": {
+    "less": "^2.5.1",
+    "less-loader": "^2.2.1"
+  },
+  "devDependencies": {
+    "babel-eslint": "~6.1.2",
+    "eslint": "~3.0.1",
+    "eslint-config-airbnb-base": "~4.0.0",
+    "eslint-plugin-babel": "~3.3.0",
+    "eslint-plugin-import": "~1.10.2"
+  },
+  "roc": {
+    "packages": [
+      "./lib/index.js"
+    ]
+  }
+}
