+++ extensions/roc-plugin-style-less/src/index.js
@@ -1,0 +1,12 @@
+// eslint-disable-next-line
+export const roc = {
+    actions: [{
+        extension: 'roc-plugin-style-css',
+        hook: 'add-style',
+        description: 'Adds Less support to Webpack.',
+        action: () => () => () => () => ({
+            extensions: ['less'],
+            loaders: `${require.resolve('less-loader')}?sourceMap`,
+        }),
+    }],
+};
