--- packages/roc-plugin-style-less/README.md
@@ -1,8 +1,0 @@
-# roc-plugin-style-less
-Plugin providing Less support for Webpack.
-
-## Documentation
-- [Actions](/packages/roc-plugin-style-less/docs/Actions.md)
-- [Commands](/packages/roc-plugin-style-less/docs/Commands.md)
-- [Hooks](/packages/roc-plugin-style-less/docs/Hooks.md)
-- [Settings](/packages/roc-plugin-style-less/docs/Settings.md)
