--- packages/roc-plugin-style-less/docs/Commands.md
@@ -1,1 +1,0 @@
-No commands available.
