--- packages/roc-plugin-style-less/docs/Hooks.md
@@ -1,1 +1,0 @@
-No hooks available.
