--- packages/roc-plugin-style-less/docs/Settings.md
@@ -1,1 +1,0 @@
-No settings available.
