--- packages/roc-plugin-style-less/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "roc-plugin-style-less",
-  "version": "1.0.0-alpha.2",
-  "description": "Plugin providing Less support for Webpack in Roc",
-  "main": "lib/index.js",
-  "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint"
-  },
-  "files": [
-    "lib"
-  ],
-  "keywords": [
-    "roc",
-    "roc-plugin",
-    "webpack",
-    "less"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/rocjs/roc-plugin-style-less.git"
-  },
-  "author": "Aftonbladet",
-  "license": "MIT",
-  "dependencies": {
-    "less": "^2.5.1",
-    "less-loader": "^2.2.1"
-  },
-  "devDependencies": {
-    "babel-eslint": "~5.0.0",
-    "eslint": "~1.10.3",
-    "eslint-config-vgno": "~5.0.0"
-  }
-}
